Added Comments on StatsLibrary. Began DoorGame.
StatsLibrary has comments, and I removed the findFactorial,
findPermutations, and findCombinations methods that used regular int's
rather than strings and BigIntegers. DoorGame has been made, though I
don't know if it is correct because it seems like it should be a 50/50
if they keep or don't keep their answer. I'll update it tomorrow. Didn't
do comments on DoorGame because I was doing it fast instead of
meticulously.